Update base.html
RCW-224 - Extra spacing of background on the right side of the page on mobile view - Fixed. 

Note:
There's a chance if you open the page too early after booting up the local server. It may show the extra white spaces but after a while it won't show anymore in the website